mqtt-client/requirements.txt
Flask
paho-mqtt
python-dotenv